+++ changelog.md
@@ -12,6 +12,7 @@ Codion Change Log
 - FilteredTable search field now has a minimum size instead of columns.
 - TextFieldBuilder.DEFAULT_TEXT_FIELD_COLUMNS configuration value removed.
 - Utilities.updateComponentTreeUI() added.
+- FilteredTableConditionPanel no longer overrides updateUI(), seems unnecessary.
 ### is.codion.framework.model
 - EntitySearchModel.LIMIT renamed DEFAULT_LIMIT.
 - EntityTableModel and EntityComboBoxModel.respondToEditEvents() renamed handleEditEvents().